pyproject.toml
[project]
name = "diagnostix"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.2",
    "celery-types>=0.26.0",
    "django>=6.0.4",
    "djangorestframework>=3.16.0",
    "djangorestframework-stubs>=3.16.9",
]
